Fix docker failure when update fails.

+++ init/20_apt_update.sh
@@ -4,10 +4,16 @@
 #
 
 # Update repositories
-apt-get update
+echo "Performing updates..."
+apt-get update 2>&1 | tee /tmp/test_update
 
-# Perform Upgrade
-apt-get -y upgrade -o Dpkg::Options::="--force-confold"
+# Verify that the updates will work.
+if [ "`cat /tmp/test_update | grep 'Failed'`" = "" ]; then
+	# Perform Upgrade
+	apt-get -y upgrade -o Dpkg::Options::="--force-confold"
 
-# Clean + purge old/obsoleted packages
-apt-get -y autoremove
+	# Clean + purge old/obsoleted packages
+	apt-get -y autoremove
+else
+	echo "Warning: Unable to update!  Check Internet connection."
+fi
